requirements.txt
fastapi==0.100.0
sqlalchemy==2.0.19
uvicorn==0.23.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
